requirements.txt
numpy
matplotlib
opencv-python
scikit-image
scipy
pillow
pandas
scikit-learn